contacts_api/pyproject.toml
[tool.poetry]
name = "contacts-api"
version = "0.1.0"
description = ""
authors = ["Rai <154880397+Rai925@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.112.0"}
sqlalchemy = "^1.4.46"
psycopg2-binary = "^2.9.9"
pydantic = "^2.8.2"
alembic = "^1.13.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-jwt-auth = "^0.5.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
auth = "^0.5.3"
bcrypt = "^4.2.0"
libgravatar = "^1.0.4"
fastapi-mail = "^1.4.1"
cloudinary = "^1.41.0"
slowapi = "^0.1.9"
redis = "^5.0.8"
pytest = "^8.3.2"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"
sphinx = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
